+++ frontend/src/components/Locality/Tabs/LithologyTab.tsx
@@ -1,6 +1,7 @@
 import { Editable, LocalityDetails, SedimentaryStructure } from '@/backendTypes'
 import { EditableTable } from '@/components/DetailView/common/EditableTable'
-import { ArrayFrame, HalfFrames, EditingModal, Grouped } from '@/components/DetailView/common/FormComponents'
+import { EditingModal } from '@/components/DetailView/common/EditingModal'
+import { ArrayFrame, HalfFrames, Grouped } from '@/components/DetailView/common/tabLayoutHelpers'
 import { useDetailContext } from '@/components/DetailView/hooks'
 import { Box, TextField } from '@mui/material'
 import { MRT_ColumnDef } from 'material-react-table'
