+++ frontend/src/components/Species/Tabs/SizeTab.tsx
@@ -1,5 +1,5 @@
 import { SpeciesDetails } from '@/backendTypes'
-import { ArrayFrame, HalfFrames } from '@components/DetailView/common/FormComponents'
+import { ArrayFrame, HalfFrames } from '@/components/DetailView/common/tabLayoutHelpers'
 import { useDetailContext } from '@components/DetailView/hooks'
 
 export const SizeTab = () => {
